requirements.txt
discord.py>=2.3.0
slack-bolt>=1.18.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
